+++ src/client/java/com/stashwalker/features/impl/NewChunksFeatureImpl.java
@@ -24,7 +24,7 @@ import net.minecraft.world.chunk.Chunk;
 
 public class NewChunksFeatureImpl extends AbstractBaseFeature implements ChunkProcessor, Renderable  {
 
-    // This List implementation is optimized for fast interation and slow add and remove, so rendering will be fast
+    // This List implementation is optimized for fast iteration and slow add and remove, so rendering will be fast
     private final List<ChunkPos> buffer = new CopyOnWriteArrayList<>();
     private final String newChunksColorKey = "newChunksColor";
     private final Color newChunksColorDefaultValue = Color.RED;
@@ -63,10 +63,7 @@ public class NewChunksFeatureImpl extends AbstractBaseFeature implements ChunkPr
 
         if (this.enabled) {
 
-            if (this.buffer.contains(chunk.getPos())) {
-
-                    buffer.remove(chunk.getPos());
-            }
+            buffer.remove(chunk.getPos());
         }
     }