+++ src/client/java/com/stashwalker/utils/FinderUtil.java
@@ -7,10 +7,13 @@ import net.minecraft.world.chunk.Chunk;
 import net.minecraft.world.chunk.ChunkStatus;
 
 import java.lang.Class;
+import java.util.ArrayList;
+import java.util.HashSet;
 import java.util.List;
-
+import java.util.Set;
+import java.util.function.Function;
 import com.stashwalker.constants.Constants;
-
+import com.stashwalker.containers.KDTree;
 import net.minecraft.entity.Entity;
 import net.minecraft.util.math.Box;
 
@@ -60,5 +63,33 @@ public class FinderUtil {
     public static boolean isBlockType (BlockPos blockPos, Block block) {
 
         return Constants.MC_CLIENT_INSTANCE.world.getBlockState(blockPos).isOf(block);
+    }
+
+    public static<T> List<T> findCloseProximityBlockPositionObjects (
+            List<T> positionObjects,
+            KDTree<T> kdTree,
+            Function<T, BlockPos> positionExtractor,
+            int amount,
+            int proximity
+    ) {
+
+        Set<T> closeProximityPositionObjects = new HashSet<>();
+        for (T current: positionObjects) {
+
+            if (closeProximityPositionObjects.contains(current)) {
+
+                continue;
+            }
+
+            Set<T> nearbyPositionObjects =
+                kdTree.rangeSearch(positionExtractor.apply(current), proximity);
+
+            if (nearbyPositionObjects.size() >= amount) {
+
+                closeProximityPositionObjects.addAll(nearbyPositionObjects);
+            }
+        }
+
+        return new ArrayList<>(closeProximityPositionObjects);
     }
 }